requirements.txt
langchain==0.3.18
langchain-community==0.3.17
langchain-openai==0.3.6
python-dotenv==1.0.1
python-telegram-bot==21.10